+++ build.sh
@@ -41,13 +41,13 @@ if [ ! -f "config/ssl/fullchain.pem" ]; then
   cp -v "config/ssl/cert.pem" "config/ssl/fullchain.pem"
 fi
 
-docker-compose build --progress plain --no-cache
+docker-compose build --progress plain $BUILD_CACHE
 
 # Run config test
-docker run -it --rm --env-file .env -v "$(pwd)/webroot:/var/www/html:ro,z" compilenix/nginx:${NGINX_VERSION} /usr/bin/nginx -t
+docker run -i --rm --env-file .env -v "$(pwd)/webroot:/var/www/html:ro,z" compilenix/nginx:${NGINX_VERSION} /usr/bin/nginx -t
 
 # Get build version info
-docker run -it --rm --env-file .env -e NGINX_ENTRYPOINT_QUIET_LOGS=y -v "$(pwd)/webroot:/var/www/html:ro,z" compilenix/nginx:${NGINX_VERSION} /usr/bin/nginx -V
+docker run -i --rm --env-file .env -e ENTRYPOINT_QUIET=y -v "$(pwd)/webroot:/var/www/html:ro,z" compilenix/nginx:${NGINX_VERSION} /usr/bin/nginx -V
 
 build_date_end_timestamp=$(date +%s)
 build_date_end_pretty=$(LC_TIME="en_US.UTF-8" TZ="GMT" date "+%a, %d %b %Y %T %Z")